% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{align_sequences}
\alias{align_sequences}
\title{Align Multiple Sequences}
\usage{
align_sequences(sequences, method = "clustal", ...)
}
\arguments{
\item{sequences}{A \code{dna_sequences} or \code{protein_sequences} object.}

\item{method}{Character. \code{"clustal"} (default) or \code{"muscle"}.}

\item{...}{Extra arguments forwarded to \code{msa::msa()}.}
}
\value{
S3 object of class \code{aligned_sequences}:
\describe{
\item{alignment}{Raw alignment from \code{msa::msa()} or character matrix.}
\item{alignment_matrix}{Character matrix (sequences x positions).}
\item{method}{Alignment method.}
\item{stats}{Data frame: \code{n_seqs}, \code{n_pos}, \code{gap_pct}, \code{coverage}.}
\item{type}{Sequence type.}
}
}
\description{
Performs multiple sequence alignment using ClustalOmega or MUSCLE via the
\strong{msa} package when available, otherwise pads to maximum length.
}
\examples{
fasta_lines <- c(
  ">Seq1", "ATGCATGCATGCATGCATGC",
  ">Seq2", "ATGCTTGCATGCATGCTTGC",
  ">Seq3", "ATGCATGCTTGCATGCATGC"
)
tmp <- tempfile(fileext = ".fasta")
writeLines(fasta_lines, tmp)
seqs <- read_fasta(tmp)
aln  <- align_sequences(seqs)
dim(aln$alignment_matrix)

}
\references{
Sievers & Higgins (2018). Clustal Omega for making accurate alignments of
many protein sequences. \emph{Protein Science}, 27(1), 135-145.
}
