% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{build_tree_ml}
\alias{build_tree_ml}
\title{Construct a Maximum Likelihood Phylogenetic Tree}
\usage{
build_tree_ml(alignment, model = NULL, optimize = TRUE, ...)
}
\arguments{
\item{alignment}{An \code{aligned_sequences} object.}

\item{model}{Character. Substitution model (default \code{"GTR"} for DNA,
\code{"LG"} for protein).}

\item{optimize}{Logical. Run \code{phangorn::optim.pml()}? Default \code{TRUE}.}

\item{...}{Additional arguments to \code{phangorn::pml()}.}
}
\value{
S3 object of class \code{phylo_tree_ml}:
\describe{
\item{tree}{\code{ape::phylo} object.}
\item{likelihood}{Log-likelihood value.}
\item{model}{Substitution model.}
\item{edge_lengths}{Named numeric branch lengths.}
\item{pml_fit}{Full \code{pml} object from phangorn.}
\item{metadata}{List: \code{n_seqs}, \code{n_positions}, \code{alignment_method}.}
}
}
\description{
Builds an ML tree from aligned sequences using \strong{phangorn}, starting from
a Neighbor-Joining tree and optimising likelihood.
}
\examples{
fasta_lines <- c(
  ">Seq1", "ATGCATGCATGCATGCATGCATGCATGC",
  ">Seq2", "ATGCTTGCATGCATGCTTGCATGCTTGC",
  ">Seq3", "ATGCATGCTTGCATGCATGCTTGCATGC",
  ">Seq4", "ATGCATGCATGCTTGCATGCATGCATGC"
)
tmp <- tempfile(fileext = ".fasta")
writeLines(fasta_lines, tmp)
seqs <- read_fasta(tmp)
aln  <- align_sequences(seqs)
tree <- build_tree_ml(aln)
print(tree)

}
\references{
Schliep (2011). phangorn: phylogenetic analysis in R.
\emph{Bioinformatics}, 27(4), 592-593.
}
