% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\docType{package}
\name{phylogenetics}
\alias{phylogenetics}
\title{Module 5: Phylogenetics & Evolutionary Mapping}
\description{
Functions for phylogenetic analysis of DNA and protein sequence data,
including FASTA parsing, multiple sequence alignment, tree construction
(Maximum Likelihood and Neighbor-Joining), bootstrap resampling,
tree visualization, and export to standard formats (Newick/Nexus).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lunganlungpg/agristat}
  \item Report bugs at \url{https://github.com/lunganlungpg/agristat/issues}
}

}
\author{
\strong{Maintainer}: Lung An Lung Pg \email{lunganlungpg@example.com}

}
\keyword{internal}
