% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{plot_tree_unrooted}
\alias{plot_tree_unrooted}
\title{Plot an Unrooted (Radial) Phylogenetic Tree}
\usage{
plot_tree_unrooted(tree, main = "Unrooted Tree", tip_labels = TRUE, ...)
}
\arguments{
\item{tree}{A \code{phylo_tree_ml}, \code{phylo_tree_nj}, \code{bootstrap_result}, or
\code{ape::phylo} object.}

\item{main}{Character. Plot title. Default \code{"Unrooted Tree"}.}

\item{tip_labels}{Logical. Show tip labels? Default \code{TRUE}.}

\item{...}{Extra arguments forwarded to \code{ape::plot.phylo()}.}
}
\value{
ggplot2 object (ggtree) or \code{NULL} invisibly (base graphics).
}
\description{
Displays a tree in unrooted/radial layout using \code{ape} or ggtree.
}
\examples{
fasta_lines <- c(
  ">Seq1", "ATGCATGCATGCATGCATGC",
  ">Seq2", "ATGCTTGCATGCATGCTTGC",
  ">Seq3", "ATGCATGCTTGCATGCATGC",
  ">Seq4", "ATGCATGCATGCTTGCATGC"
)
tmp <- tempfile(fileext = ".fasta")
writeLines(fasta_lines, tmp)
seqs <- read_fasta(tmp)
aln  <- align_sequences(seqs)
tree <- build_tree_nj(aln)
plot_tree_unrooted(tree)

}
