% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-layouts.R
\name{print.agristat_aov}
\alias{print.agristat_aov}
\title{Print an agristat_aov Object}
\usage{
\method{print}{agristat_aov}(x, ...)
}
\arguments{
\item{x}{An \code{agristat_aov} object.}

\item{...}{Passed to the underlying model's print method.}
}
\value{
\code{x} invisibly.
}
\description{
Print an agristat_aov Object
}
