% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{print.assumption_report}
\alias{print.assumption_report}
\title{Print an \code{assumption_report} Object}
\usage{
\method{print}{assumption_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"assumption_report"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print an \code{assumption_report} Object
}
