% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{print.auto_transform_result}
\alias{print.auto_transform_result}
\title{Print an \code{auto_transform_result} Object}
\usage{
\method{print}{auto_transform_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"auto_transform_result"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print an \code{auto_transform_result} Object
}
