% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{print.chi_square_result}
\alias{print.chi_square_result}
\title{Print a \code{chi_square_result} Object}
\usage{
\method{print}{chi_square_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"chi_square_result"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a \code{chi_square_result} Object
}
