% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-layouts.R
\name{print.field_design}
\alias{print.field_design}
\title{Print a field_design Object}
\usage{
\method{print}{field_design}(x, ...)
}
\arguments{
\item{x}{A \code{field_design} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a field_design Object
}
