% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{reroot_tree}
\alias{reroot_tree}
\title{Reroot a Phylogenetic Tree}
\usage{
reroot_tree(tree, outgroup)
}
\arguments{
\item{tree}{A \code{phylo_tree_ml}, \code{phylo_tree_nj}, \code{bootstrap_result}, or
\code{ape::phylo} object.}

\item{outgroup}{Character vector of one or more tip names, \strong{or} an
integer node number.}
}
\value{
A rerooted \code{ape::phylo} object.
}
\description{
Reroots a tree on a specified outgroup taxon or internal node.
}
\examples{
fasta_lines <- c(
  ">Seq1", "ATGCATGCATGCATGCATGC",
  ">Seq2", "ATGCTTGCATGCATGCTTGC",
  ">Seq3", "ATGCATGCTTGCATGCATGC",
  ">Seq4", "ATGCATGCATGCTTGCATGC"
)
tmp <- tempfile(fileext = ".fasta")
writeLines(fasta_lines, tmp)
seqs <- read_fasta(tmp)
aln  <- align_sequences(seqs)
tree <- build_tree_nj(aln)
rerooted <- reroot_tree(tree, outgroup = "Seq4")
rerooted$tip.label

}
