% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{suggest_nonparametric}
\alias{suggest_nonparametric}
\title{Suggest Non-Parametric Alternatives}
\usage{
suggest_nonparametric(assumption_report, n_groups = NULL)
}
\arguments{
\item{assumption_report}{An object of class \code{"assumption_report"} produced
by \code{\link[=check_assumptions]{check_assumptions()}}.}

\item{n_groups}{Integer.  Number of treatment groups.  If \code{NULL}
(default), the value stored in \code{assumption_report$n_groups} is used.}
}
\value{
A named character vector of recommendations.  Names are
\code{"normality"} and/or \code{"homogeneity"} for the relevant violations.
When all assumptions pass, a single \code{"parametric"} element is returned.
}
\description{
Examines an \code{assumption_report} and recommends non-parametric (or
variance-robust) alternatives when parametric assumptions are violated.
}
\examples{
set.seed(99)
x <- c(rexp(20), rnorm(20, 10))   # non-normal
grp <- rep(c("A", "B"), each = 20)
report <- check_assumptions(x, groups = grp)
suggest_nonparametric(report)

}
\seealso{
\code{\link[=check_assumptions]{check_assumptions()}}, \code{\link[=auto_transform]{auto_transform()}}
}
