% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{summary.assumption_report}
\alias{summary.assumption_report}
\title{Summarise an \code{assumption_report} Object}
\usage{
\method{summary}{assumption_report}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"assumption_report"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Returns \code{object} invisibly.
}
\description{
Summarise an \code{assumption_report} Object
}
