% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion.R
\name{summary.auto_transform_result}
\alias{summary.auto_transform_result}
\title{Summarise an \code{auto_transform_result} Object}
\usage{
\method{summary}{auto_transform_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"auto_transform_result"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Returns \code{object} invisibly.
}
\description{
Summarise an \code{auto_transform_result} Object
}
