% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-layouts.R
\name{summary.field_design}
\alias{summary.field_design}
\title{Summarise a field_design Object}
\usage{
\method{summary}{field_design}(object, ...)
}
\arguments{
\item{object}{A \code{field_design} object.}

\item{...}{Ignored.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a field_design Object
}
